+++ zubhub_frontend/zubhub/src/views/profile/profileScripts.js
@@ -7,16 +7,16 @@ import { getUserDrafts } from '../../store/actions/projectActions';
  *
  * @todo - describe function's signature
  */ export const getUserProfile = props => {
-  // let username = props.match.params.username;
+  let username = props.match.params.username;
 
-  // if (!username) {
-  //   username = props.auth.username;
-  // } else if (props.auth.username === username) props.history.replace('/profile');
-  // return props.getUserProfile({
-  //   username,
-  //   token: props.auth.token,
-  //   t: props.t,
-  // });
+  if (!username) {
+    username = props.auth.username;
+  } else if (props.auth.username === username) props.history.replace('/profile');
+  return props.getUserProfile({
+    username,
+    token: props.auth.token,
+    t: props.t,
+  });
 };
 
 /**
@@ -26,19 +26,19 @@ import { getUserDrafts } from '../../store/actions/projectActions';
  * @todo - describe function's signature
  */
 export const copyProfileUrl = (profile, props, toast) => {
-  // const tempInput = document.createElement('textarea');
-  // tempInput.value = `${document.location.origin}/creators/${profile.username}`;
-  // tempInput.style.top = '0';
-  // tempInput.style.top = '0';
-  // tempInput.style.position = 'fixed';
-  // const rootElem = document.querySelector('#root');
-  // rootElem.appendChild(tempInput);
-  // tempInput.focus();
-  // tempInput.select();
-  // if (document.execCommand('copy')) {
-  //   toast.success(props.t('profile.toastSuccess'));
-  //   rootElem.removeChild(tempInput);
-  // }
+  const tempInput = document.createElement('textarea');
+  tempInput.value = `${document.location.origin}/creators/${profile.username}`;
+  tempInput.style.top = '0';
+  tempInput.style.top = '0';
+  tempInput.style.position = 'fixed';
+  const rootElem = document.querySelector('#root');
+  rootElem.appendChild(tempInput);
+  tempInput.focus();
+  tempInput.select();
+  if (document.execCommand('copy')) {
+    toast.success(props.t('profile.toastSuccess'));
+    rootElem.removeChild(tempInput);
+  }
 };
 
 /**
@@ -48,34 +48,34 @@ export const copyProfileUrl = (profile, props, toast) => {
  * @todo - describe function's signature
  */
 export const updateProjects = (res, projects, props, toast) => {
-  // return res
-  //   .then(res => {
-  //     if (res.project && res.project.title) {
-  //       if (projects[0]) {
-  //         projects = projects.map(project =>
-  //           project.id === res.project.id ? res.project : project,
-  //         );
-  //       } else {
-  //         projects = projects.results.map(project =>
-  //           project.id === res.project.id ? res.project : project,
-  //         );
-  //       }
-  //       return { results: projects };
-  //     } else {
-  //       res = Object.keys(res)
-  //         .map(key => res[key])
-  //         .join('\n');
-  //       throw new Error(res);
-  //     }
-  //   })
-  //   .catch(error => {
-  //     if (error.message.startsWith('Unexpected')) {
-  //       toast.warning(props.t('profile.errors.unexpected'));
-  //     } else {
-  //       toast.warning(error.message);
-  //     }
-  //     return { loading: false };
-  //   });
+  return res
+    .then(res => {
+      if (res.project && res.project.title) {
+        if (projects[0]) {
+          projects = projects.map(project =>
+            project.id === res.project.id ? res.project : project,
+          );
+        } else {
+          projects = projects.results.map(project =>
+            project.id === res.project.id ? res.project : project,
+          );
+        }
+        return { results: projects };
+      } else {
+        res = Object.keys(res)
+          .map(key => res[key])
+          .join('\n');
+        throw new Error(res);
+      }
+    })
+    .catch(error => {
+      if (error.message.startsWith('Unexpected')) {
+        toast.warning(props.t('profile.errors.unexpected'));
+      } else {
+        toast.warning(error.message);
+      }
+      return { loading: false };
+    });
 };
 
 /**
@@ -85,28 +85,28 @@ export const updateProjects = (res, projects, props, toast) => {
  * @todo - describe function's signature
  */
 export const updateDrafts = (res, projects, props, toast) => {
-  // return res
-  //   .then(res => {
-  //     if (res.project && res.project.title) {
-  //       projects = projects.map(project =>
-  //         project.id === res.project.id ? res.project : project,
-  //       );
-  //       return { drafts: projects };
-  //     } else {
-  //       res = Object.keys(res)
-  //         .map(key => res[key])
-  //         .join('\n');
-  //       throw new Error(res);
-  //     }
-  //   })
-  //   .catch(error => {
-  //     if (error.message.startsWith('Unexpected')) {
-  //       toast.warning(props.t('profile.errors.unexpected'));
-  //     } else {
-  //       toast.warning(error.message);
-  //     }
-  //     return { loading: false };
-  //   });
+  return res
+    .then(res => {
+      if (res.project && res.project.title) {
+        projects = projects.map(project =>
+          project.id === res.project.id ? res.project : project,
+        );
+        return { drafts: projects };
+      } else {
+        res = Object.keys(res)
+          .map(key => res[key])
+          .join('\n');
+        throw new Error(res);
+      }
+    })
+    .catch(error => {
+      if (error.message.startsWith('Unexpected')) {
+        toast.warning(props.t('profile.errors.unexpected'));
+      } else {
+        toast.warning(error.message);
+      }
+      return { loading: false };
+    });
 };
 
 /**
@@ -116,11 +116,11 @@ export const updateDrafts = (res, projects, props, toast) => {
  * @todo - describe function's signature
  */
 export const toggleFollow = (id, props) => {
-  // if (!props.auth.token) {
-  //   props.history.push('/login');
-  // } else {
-  //   return props.toggleFollow({ id, token: props.auth.token, t: props.t });
-  // }
+  if (!props.auth.token) {
+    props.history.push('/login');
+  } else {
+    return props.toggleFollow({ id, token: props.auth.token, t: props.t });
+  }
 };
 
 /**
@@ -137,27 +137,27 @@ export const toggleFollow = (id, props) => {
  * @returns {Array} - sorted array of tags.
  */
 export const sortTags = tags => {
-  // return tags.sort((a, b) => {
-  //   if (BASE_TAGS.includes(a) && BASE_TAGS.includes(b)) {
-  //     //standard reverse sort.
-  //     if (a > b) {
-  //       return -1;
-  //     } else {
-  //       return 1;
-  //     }
-  //   } else if (BASE_TAGS.includes(a)) {
-  //     return -1;
-  //   } else if (BASE_TAGS.includes(b)) {
-  //     return 1;
-  //   } else {
-  //     //standard sort.
-  //     if (a > b) {
-  //       return 1;
-  //     } else {
-  //       return -1;
-  //     }
-  //   }
-  // });
+  return tags.sort((a, b) => {
+    if (BASE_TAGS.includes(a) && BASE_TAGS.includes(b)) {
+      //standard reverse sort.
+      if (a > b) {
+        return -1;
+      } else {
+        return 1;
+      }
+    } else if (BASE_TAGS.includes(a)) {
+      return -1;
+    } else if (BASE_TAGS.includes(b)) {
+      return 1;
+    } else {
+      //standard sort.
+      if (a > b) {
+        return 1;
+      } else {
+        return -1;
+      }
+    }
+  });
 };
 
 /**
@@ -177,7 +177,7 @@ export const handleMoreMenuOpen = e => {
  * @todo - describe function's signature
  */
 export const handleMoreMenuClose = () => {
-  // return { more_anchor_el: null };
+  return { more_anchor_el: null };
 };
 
 /**
@@ -187,8 +187,8 @@ export const handleMoreMenuClose = () => {
  * @todo - describe function's signature
  */
 export const handleToggleDeleteAccountModal = state => {
-  // const open_delete_account_modal = !state.open_delete_account_modal;
-  // return { open_delete_account_modal, more_anchor_el: null };
+  const open_delete_account_modal = !state.open_delete_account_modal;
+  return { open_delete_account_modal, more_anchor_el: null };
 };
 
 /**
@@ -198,14 +198,14 @@ export const handleToggleDeleteAccountModal = state => {
  * @todo - describe function's signature
  */
 export const deleteAccount = (username_el, props) => {
-  // if (username_el.current.firstChild.value !== props.auth.username) {
-  //   return { dialogError: props.t('profile.delete.errors.incorrectUsernme') };
-  // } else {
-  //   return props.deleteAccount({
-  //     token: props.auth.token,
-  //     history: props.history,
-  //     logout: props.logout,
-  //     t: props.t,
-  //   });
-  // }
+  if (username_el.current.firstChild.value !== props.auth.username) {
+    return { dialogError: props.t('profile.delete.errors.incorrectUsernme') };
+  } else {
+    return props.deleteAccount({
+      token: props.auth.token,
+      history: props.history,
+      logout: props.logout,
+      t: props.t,
+    });
+  }
 };
